The work that was done for 1.14.x
Some progress was made figuring out how to update to 1.14, but I hit a wall trying to get models to refresh and never came back to it. This might be a useful reference on some of the stuff that likely needs to to change to bring BiblioCraft up to date.