Fixed some issues that came up during testing
My quick fix to get the decided server running caused some issues since the packets still needed to be loaded server side, but several of the client ones were using PlayerEntity from client side that would error out on the dedicated server. So instead of loading the packets just on the client side, they are back with the server, but the methods that use the EntityPlayer is now called from a new function that is client side only, that seems to have fixed several packet related issues on dedicated server.